--- a/mlflow/entities/run_info.py
+++ b/mlflow/entities/run_info.py
@@ -110,7 +110,7 @@ class RunInfo(_MLflowObject):
         """Start time of the run, in number of milliseconds since the UNIX epoch."""
         return self._start_time
 
-    @property
+    @orderable_attribute
     def end_time(self):
         """End time of the run, in number of milliseconds since the UNIX epoch."""
         return self._end_time