… tests/store/tracking/test_sqlalchemy_store.py